+++ android/src/main/kotlin/com/oguzhnatly/flutter_android_auto/FAAEnums.kt
@@ -19,4 +19,6 @@ enum class FAAChannelTypes {
     onListSectionSelected,
     onToggleCheckedChange,
     onScreenBackButtonPressed,
+    updateMessageTemplate,
+    updateLongMessageTemplate,
 }