+++ lib/aa_models/all.dart
@@ -1,2 +1,3 @@
 export 'list/all.dart';
+export 'message/all.dart';
 export 'template.dart';