+++ lib/android_auto_worker.dart
@@ -43,47 +43,50 @@ class FlutterAndroidAuto {
     _eventBroadcast = _androidAutoController.eventChannel
         .receiveBroadcastStream()
         .listen((event) {
-      final FAAChannelTypes receivedChannelType =
-          EnumUtils.enumFromString(FAAChannelTypes.values, event['type']);
+          final FAAChannelTypes receivedChannelType = EnumUtils.enumFromString(
+            FAAChannelTypes.values,
+            event['type'],
+          );
 
-      switch (receivedChannelType) {
-        case FAAChannelTypes.onAndroidAutoConnectionChange:
-          final ConnectionStatusTypes connectionStatus =
-              EnumUtils.enumFromString(
-            ConnectionStatusTypes.values,
-            event['data']['status'],
-          );
-          _connectionStatus = connectionStatus.name;
-          if (_onAndroidAutoConnectionChange != null) {
-            _onAndroidAutoConnectionChange!(connectionStatus);
+          switch (receivedChannelType) {
+            case FAAChannelTypes.onAndroidAutoConnectionChange:
+              final ConnectionStatusTypes connectionStatus =
+                  EnumUtils.enumFromString(
+                    ConnectionStatusTypes.values,
+                    event['data']['status'],
+                  );
+              _connectionStatus = connectionStatus.name;
+              if (_onAndroidAutoConnectionChange != null) {
+                _onAndroidAutoConnectionChange!(connectionStatus);
+              }
+              break;
+            case FAAChannelTypes.onListItemSelected:
+              _androidAutoController.processFAAListItemSelectedChannel(
+                event['data']['elementId'],
+              );
+              break;
+            case FAAChannelTypes.onListSectionSelected:
+              _androidAutoController.processFAAListSectionSelectedChannel(
+                event['data']['elementId'],
+                event['data']['selectedIndex'],
+              );
+              break;
+            case FAAChannelTypes.onToggleCheckedChange:
+              _androidAutoController.processFAAToggleCheckedChangeChannel(
+                event['data']['elementId'],
+                event['data']['checked'],
+              );
+              break;
+            case FAAChannelTypes.onScreenBackButtonPressed:
+              FlutterAndroidAutoController.templateHistory.removeWhere(
+                (AATemplate item) =>
+                    item.uniqueId == event['data']['elementId'],
+              );
+              break;
+            default:
+              break;
           }
-          break;
-        case FAAChannelTypes.onListItemSelected:
-          _androidAutoController.processFAAListItemSelectedChannel(
-            event['data']['elementId'],
-          );
-          break;
-        case FAAChannelTypes.onListSectionSelected:
-          _androidAutoController.processFAAListSectionSelectedChannel(
-            event['data']['elementId'],
-            event['data']['selectedIndex'],
-          );
-          break;
-        case FAAChannelTypes.onToggleCheckedChange:
-          _androidAutoController.processFAAToggleCheckedChangeChannel(
-            event['data']['elementId'],
-            event['data']['checked'],
-          );
-          break;
-        case FAAChannelTypes.onScreenBackButtonPressed:
-          FlutterAndroidAutoController.templateHistory.removeWhere(
-            (AATemplate item) => item.uniqueId == event['data']['elementId'],
-          );
-          break;
-        default:
-          break;
-      }
-    });
+        });
   }
 
   /// A function that will disconnect all event listeners from Android Auto. The action
@@ -127,14 +130,12 @@ class FlutterAndroidAuto {
     return _connectionStatus;
   }
 
-  static Future<void> setRootTemplate({
-    required AATemplate template,
-  }) async {
+  static Future<void> setRootTemplate({required AATemplate template}) async {
     final bool? isCompleted = await _androidAutoController
         .flutterToNativeModule(FAAChannelTypes.setRootTemplate, {
-      'template': template.toJson(),
-      'runtimeType': _getAARuntimeTypeString(template),
-    });
+          'template': template.toJson(),
+          'runtimeType': _getAARuntimeTypeString(template),
+        });
 
     if (isCompleted == true) {
       if (FlutterAndroidAutoController.templateHistory.isEmpty) {
@@ -175,7 +176,7 @@ class FlutterAndroidAuto {
   /// - If animated is true, CarPlay animates the presentation of the template.
   ///
   /// [!] CarPlay can only present one modal template at a time.
-/*static Future<void> showAlert({
+  /*static Future<void> showAlert({
     required CPAlertTemplate template,
     bool animated = true,
   }) {
@@ -199,7 +200,7 @@ class FlutterAndroidAuto {
   /// - If animated is true, CarPlay animates the presentation of the template.
   ///
   /// [!] CarPlay can only present one modal template at a time.
-/* static Future<void> showActionSheet({
+  /* static Future<void> showActionSheet({
     required CPActionSheetTemplate template,
     bool animated = true,
   }) {
@@ -220,10 +221,8 @@ class FlutterAndroidAuto {
   ///
   /// The history will be updated accordingly to the screen lifecycle.
   static Future<bool> pop() async {
-    final bool? isCompleted =
-        await _androidAutoController.flutterToNativeModule(
-      FAAChannelTypes.popTemplate,
-    );
+    final bool? isCompleted = await _androidAutoController
+        .flutterToNativeModule(FAAChannelTypes.popTemplate);
 
     return isCompleted ?? false;
   }
@@ -232,16 +231,14 @@ class FlutterAndroidAuto {
   ///
   /// The history will be updated accordingly to the screen lifecycle.
   static Future<bool> popToRoot() async {
-    final bool? isCompleted =
-        await _androidAutoController.flutterToNativeModule(
-      FAAChannelTypes.popToRootTemplate,
-    );
+    final bool? isCompleted = await _androidAutoController
+        .flutterToNativeModule(FAAChannelTypes.popToRootTemplate);
     return isCompleted ?? false;
   }
 
   /// Removes a modal template. Since [CPAlertTemplate] and [CPActionSheetTemplate] are both
   /// modals, they can be removed. If animated is true, CarPlay animates the transition between templates.
-/* static Future<bool> popModal({bool animated = true}) async {
+  /* static Future<bool> popModal({bool animated = true}) async {
     FlutterCarPlayController.currentPresentTemplate = null;
     return await _androidAutoController.flutterToNativeModule(
       FAAChannelTypes.closePresent,
@@ -252,18 +249,16 @@ class FlutterAndroidAuto {
   /// Adds a template to the navigation hierarchy and displays it.
   ///
   /// - template is to add to the navigation hierarchy. **Must be one of the type:**
-  /// [AAGridTemplate] or [AAListTemplate] [AAInformationTemplat] [AAPointOfInterestTemplate]
+  /// [AAGridTemplate] or [AAListTemplate] [AAInformationTemplat] [AAPointOfInterestTemplate] [AAMessageTemplate] [AALongMessageTemplate]
   ///
   /// Be aware of the restrictions : https://developer.android.com/training/cars/apps#template-restrictions
   /// - Max 5 templates in the navigation stack.
-  static Future<bool> push({
-    required AATemplate template,
-  }) async {
+  static Future<bool> push({required AATemplate template}) async {
     final bool? isCompleted = await _androidAutoController
         .flutterToNativeModule(FAAChannelTypes.pushTemplate, <String, dynamic>{
-      'template': template.toJson(),
-      'runtimeType': _getAARuntimeTypeString(template),
-    });
+          'template': template.toJson(),
+          'runtimeType': _getAARuntimeTypeString(template),
+        });
     if (isCompleted == true) {
       FlutterAndroidAutoController.templateHistory.add(template);
     }
@@ -278,6 +273,8 @@ class FlutterAndroidAuto {
   /// Uses explicit type checks to ensure compatibility with Dart obfuscation.
   static String _getAARuntimeTypeString(AATemplate template) {
     if (template is AAListTemplate) return 'FAAListTemplate';
+    if (template is AAMessageTemplate) return 'FAAMessageTemplate';
+    if (template is AALongMessageTemplate) return 'FAALongMessageTemplate';
     return 'FAA${template.runtimeType}';
   }
 }