+++ lib/constants/private_constants.dart
@@ -48,8 +48,9 @@ enum FAAChannelTypes {
   updateListTemplateSections,
   onListItemSelected,
   onListItemSelectedComplete,
-  updateListTemplateSections,
   onListSectionSelected,
   onToggleCheckedChange,
   onScreenBackButtonPressed,
+  updateMessageTemplate,
+  updateLongMessageTemplate,
 }