+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Next
+
+- Add Android Auto `AAMessageTemplate` support with update APIs.
+- Add Android Auto `AALongMessageTemplate` support with update APIs.
+
 ## 1.3.3
 
 - Expose `onPop` callbacks on pushed CarPlay templates (#106) (ty @sINFdorako)